Add project-wide search.
As noted in the documents, project-wide search is now implemented. This
is the last major planned feature for vim-fixme, and as such earves as
something of a beta release. I will start using it myself for a while,
and see if there are any outstanding issues.

Main unresilved issues at this time are:
  - Need a better README
  - The single file search is currently th easier mapping (2 keys as
  opposed to a key and a chord.) Is this desirable? Is it truly used
  more often?